+++ packages/protocol/src/enums/input-lock-flags.ts
@@ -1,0 +1,7 @@
+enum InputLockFlags {
+	None = 0x0,
+	Camera = 0x2,
+	Movement = 0x4
+}
+
+export { InputLockFlags };