+++ packages/protocol/src/proto/data/index.ts
@@ -64,6 +64,7 @@ export * from "./set-local-player-as-initialized";
 export * from "./network-chunk-publisher-update";
 export * from "./biome-definition-list";
 export * from "./level-sound-event";
+export * from "./on-screen-texture-animation";
 export * from "./emote";
 export * from "./network-settings";
 export * from "./player-auth-input";
@@ -114,3 +115,5 @@ export * from "./block-event";
 export * from "./entity-pick-request";
 export * from "./hurt-armor";
 export * from "./show-credits";
+
+export * from "./update-client-input-locks";