+++ packages/protocol/src/proto/data/packets.ts
@@ -111,6 +111,8 @@ import { BlockEventPacket } from "./block-event";
 import { EntityPickRequestPacket } from "./entity-pick-request";
 import { HurtArmorPacket } from "./hurt-armor";
 import { ShowCreditsPacket } from "./show-credits";
+import { UpdateClientInputLocksPacket } from "./update-client-input-locks";
+import { OnScreenTextureAnimationPacket } from "./on-screen-texture-animation";
 
 const Packets = {
 	[Packet.Login]: LoginPacket, // 1
@@ -192,6 +194,7 @@ const Packets = {
 	[Packet.NetworkChunkPublisherUpdate]: NetworkChunkPublisherUpdatePacket, // 121
 	[Packet.BiomeDefinitionList]: BiomeDefinitionListPacket, // 122
 	[Packet.LevelSoundEvent]: LevelSoundEventPacket, // 123
+	[Packet.OnScreenTextureAnimation]: OnScreenTextureAnimationPacket, // 130
 	[Packet.Emote]: EmotePacket, // 138
 	[Packet.CompletedUsingItem]: CompletedUsingItemPacket, // 142
 	[Packet.NetworkSettings]: NetworkSettingsPacket, // 143
@@ -214,6 +217,7 @@ const Packets = {
 	[Packet.UpdateAdventureSettings]: UpdateAdventureSettingsPacket, // 188
 	[Packet.DeathInfo]: DeathInfoPacket, // 189
 	[Packet.RequestNetworkSettings]: RequestNetworkSettingsPacket, // 193
+	[Packet.UpdateClientInputLocks]: UpdateClientInputLocksPacket, // 196
 	[Packet.CameraPresetsPacket]: CameraPresetsPacket, // 198
 	[Packet.CameraInstructions]: CameraInstructionsPacket, // 300
 	[Packet.OpenSign]: OpenSignPacket, // 303