+++ packages/world/src/components/entity/flag/flag.ts
@@ -24,7 +24,7 @@ class EntityFlagComponent extends EntityComponent {
 		const flag = this.entity.flags.get(this.flag);
 
 		// Check if the data was not found
-		if (!flag)
+		if (flag === undefined)
 			// Throw an error as this should never happen if the component was set up correctly
 			throw new Error(`The entity does not have the ${this.flag} flag.`);