+++ packages/world/src/components/player/player-component.ts
@@ -1,5 +1,6 @@
 import { EntityComponent } from "../entity";
 
+import type { Gamemode } from "@serenityjs/protocol";
 import type { Player } from "../../player";
 
 abstract class PlayerComponent extends EntityComponent {
@@ -18,6 +19,13 @@ abstract class PlayerComponent extends EntityComponent {
 		super(player, identifier);
 		this.player = player;
 	}
+
+	/**
+	 * Called when a player's gamemode is changed.
+	 * @param pervious The previous gamemode of the player.
+	 * @param current The current gamemode of the player.
+	 */
+	public onGamemodeChange?(pervious: Gamemode, current: Gamemode): void;
 }
 
 export { PlayerComponent };