+++ packages/entity/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @serenityjs/entity
 
+## 0.5.0
+
+### Minor Changes
+
+- eea5f6d: init v0.5.0-beta
+
+### Patch Changes
+
+- Updated dependencies [eea5f6d]
+  - @serenityjs/data@0.5.0
+
 ## 0.4.4
 
 ### Patch Changes