+++ packages/network/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @serenityjs/network
 
+## 0.5.0
+
+### Minor Changes
+
+- eea5f6d: init v0.5.0-beta
+
+### Patch Changes
+
+- Updated dependencies [eea5f6d]
+  - @serenityjs/emitter@0.5.0
+  - @serenityjs/logger@0.5.0
+  - @serenityjs/protocol@0.5.0
+  - @serenityjs/raknet@0.5.0
+
 ## 0.4.4
 
 ### Patch Changes